requirements.txt
Flask==2.0.3
gunicorn==20.1.0
PyPDF2==1.26.0
python-docx==0.8.11
numpy==1.21.2
transformers==4.11.3
torch==1.10.0
faiss-cpu==1.7.1
python-dotenv==0.19.2